+++ src/RecoveryServices/RecoveryServices.Backup.ServiceClientAdapter/BMSAPIs/ContainerAPIs.cs
@@ -23,6 +23,28 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
 {
     public partial class ServiceClientAdapter
     {
+        private static RestAzureNS.AzureOperationResponse ToAzureOperationResponse<THeader>(
+            RestAzureNS.AzureOperationHeaderResponse<THeader> response)
+        {
+            return new RestAzureNS.AzureOperationResponse
+            {
+                Request = response.Request,
+                Response = response.Response,
+                RequestId = response.RequestId
+            };
+        }
+
+        private static RestAzureNS.AzureOperationResponse<TBody> ToAzureOperationResponse<TBody, THeader>(
+            RestAzureNS.AzureOperationResponse<TBody, THeader> response)
+        {
+            return new RestAzureNS.AzureOperationResponse<TBody>
+            {
+                Body = response.Body,
+                Request = response.Request,
+                Response = response.Response,
+                RequestId = response.RequestId
+            };
+        }
 
         /// <summary>
         /// Inquire protection containers in the vault according to the query params
@@ -39,11 +61,11 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string resourceGroupName = null)
         {
             return BmsAdapter.Client.ProtectionContainers.InquireWithHttpMessagesAsync(
-                vaultName ?? BmsAdapter.GetResourceName(),
                 resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                vaultName ?? BmsAdapter.GetResourceName(),
                 AzureFabricName,
                 containerName,
-                queryFilter,
+                queryFilter?.Filter,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result;
         }
 
@@ -65,7 +87,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
                 () => BmsAdapter.Client.BackupProtectionContainers.ListWithHttpMessagesAsync(
                     vaultName ?? BmsAdapter.GetResourceName(),
                     resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                    queryFilter,
+                    queryFilter?.Filter,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             Func<string, RestAzureNS.IPage<ProtectionContainerResource>> listNextAsync =
@@ -89,8 +111,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string containerName = null)
         {
             ProtectionContainerResource container = BmsAdapter.Client.ProtectionContainers.GetWithHttpMessagesAsync(
-                    vaultName ?? BmsAdapter.GetResourceName(),
                     resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                    vaultName ?? BmsAdapter.GetResourceName(),
                     AzureFabricName,
                     containerName,                    
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
@@ -113,9 +135,9 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             queryParams.Top = 200;
             Func<RestAzureNS.IPage<BackupEngineBaseResource>> listAsync =
                 () => BmsAdapter.Client.BackupEngines.ListWithHttpMessagesAsync(
-                    vaultName ?? BmsAdapter.GetResourceName(),
                     resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                    queryParams,
+                    vaultName ?? BmsAdapter.GetResourceName(),
+                    queryParams?.Filter,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             Func<string, RestAzureNS.IPage<BackupEngineBaseResource>> listNextAsync =
@@ -143,7 +165,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
                     vaultName ?? BmsAdapter.GetResourceName(),
                     resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
                     AzureFabricName,
-                    queryFilter,
+                    queryFilter?.Filter,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             Func<string, RestAzureNS.IPage<ProtectableContainerResource>> listNextAsync =
@@ -164,7 +186,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             Func<IEnumerable<BackupManagementUsage>> listAsync = () => BmsAdapter.Client.BackupUsageSummaries.ListWithHttpMessagesAsync(
                     vaultName,
                     resourceGroupName,
-                    queryFilter,
+                    queryFilter?.Filter,
                     skipToken: null,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
             
@@ -184,7 +206,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
                 vaultName ?? BmsAdapter.GetResourceName(),
                 resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
                 AzureFabricName,
-                queryParam,
+                queryParam?.Filter,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result;
             return response;
         }
@@ -200,13 +222,13 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string resourceGroupName = null)
         {
             var response = BmsAdapter.Client.ProtectionContainers.RegisterWithHttpMessagesAsync(
-                vaultName ?? BmsAdapter.GetResourceName(),
                 resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                vaultName ?? BmsAdapter.GetResourceName(),
                 AzureFabricName,
                 containerName,
                 parameters,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result;
-            return response;
+            return ToAzureOperationResponse<ProtectionContainerResource, ProtectionContainersRegisterHeaders>(response);
         }
 
         /// <summary>
@@ -240,8 +262,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string resourceGroupName = null)
         {
             var response = BmsAdapter.Client.ProtectionContainers.UnregisterWithHttpMessagesAsync(
-            vaultName ?? BmsAdapter.GetResourceName(),
             resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+            vaultName ?? BmsAdapter.GetResourceName(),
             AzureFabricName,
             containerName,
             cancellationToken: BmsAdapter.CmdletCancellationToken).Result;